+++ Applet/AndroidSEProviderLib/src/com/android/javacard/seprovider/KMHmacKey.java
@@ -48,6 +48,6 @@ public class KMHmacKey implements KMKey {
 
   @Override
   public short getPublicKey(byte[] buf, short offset) {
-    return 0;
+    return (short) 0;
   }
 }
